+++ CSManagement/CSManagement/Controllers/Short_CourseController.cs
@@ -17,14 +17,14 @@ namespace CSManagement.Controllers
         // GET: Short_Course
         public ActionResult Index()
         {
+            if (Session["AJ"] == null) return RedirectToAction("Index", "Home");
             return View(db.Short_Course.ToList());
         }
-
-        
 
         // GET: Short_Course/Details/5
         public ActionResult Details(int? id)
         {
+            if (Session["AJ"] == null) return RedirectToAction("Index", "Home");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -40,6 +40,7 @@ namespace CSManagement.Controllers
         // GET: Short_Course/Create
         public ActionResult Create()
         {
+            if (Session["AJ"] == null) return RedirectToAction("Index", "Home");
             return View();
         }
 
@@ -64,6 +65,7 @@ namespace CSManagement.Controllers
         // GET: Short_Course/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session["AJ"] == null) return RedirectToAction("Index", "Home");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -95,6 +97,7 @@ namespace CSManagement.Controllers
         // GET: Short_Course/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (Session["AJ"] == null) return RedirectToAction("Index", "Home");
             Short_Course short_Course = db.Short_Course.Find(id);
             db.Short_Course.Remove(short_Course);
             db.SaveChanges();
